use new decorator in meshtools
