+++ felupe/mesh/_base.py
@@ -33,8 +33,9 @@ def line_line(a=0, b=1, n=2):
     "Line generator."
     points = np.linspace(a, b, n).reshape(-1, 1)
     cells = np.repeat(np.arange(n), 2)[1:-1].reshape(-1, 2)
+    cell_type = "line"
 
-    return points, cells
+    return points, cells, cell_type
 
 
 def rectangle_quad(a=(0, 0), b=(1, 1), n=(2, 2)):
@@ -46,12 +47,12 @@ def rectangle_quad(a=(0, 0), b=(1, 1), n=(2, 2)):
     if not isinstance(n, array_like):
         n = np.full(dim, n, dtype=int)
 
-    line = line_line(a[0], b[0], n[0])
+    line = line_line(a=a[0], b=b[0], n=n[0])
 
-    points, cells = expand(line, n[-1], b[-1] - a[-1])
+    points, cells, cell_type = expand(*line, n=n[-1], z=b[-1] - a[-1])
     points[:, -1] += a[-1]
 
-    return points, cells
+    return points, cells, cell_type
 
 
 def cube_hexa(a=(0, 0, 0), b=(1, 1, 1), n=(2, 2, 2)):
@@ -63,9 +64,9 @@ def cube_hexa(a=(0, 0, 0), b=(1, 1, 1), n=(2, 2, 2)):
     if not isinstance(n, array_like):
         n = np.full(dim, n, dtype=int)
 
-    rectangle = rectangle_quad(a[:-1], b[:-1], n[:-1])
+    rectangle = rectangle_quad(a=a[:-1], b=b[:-1], n=n[:-1])
 
-    points, cells = expand(rectangle, n[-1], b[-1] - a[-1])
+    points, cells, cell_type = expand(*rectangle, n=n[-1], z=b[-1] - a[-1])
     points[:, -1] += a[-1]
 
-    return points, cells
+    return points, cells, cell_type
